+++ vitest.config.ts
@@ -1,0 +1,20 @@
+/// <reference types="vitest" />
+
+import path from 'node:path'
+import react from '@vitejs/plugin-react'
+import { defineConfig } from 'vitest/config'
+
+export default defineConfig({
+  plugins: [react()],
+  test: {
+    globals: true,
+    environment: 'jsdom',
+    setupFiles: ['./testing/setup.ts'],
+    include: ['**/*.test.tsx', '**/*.test.ts'],
+  },
+  resolve: {
+    alias: {
+      '@': path.resolve(__dirname, 'src')
+    },
+  }
+})